% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenfunctions.R
\name{.clickable}
\alias{.clickable}
\title{plasmoRUtils}
\usage{
.clickable(url)
}
\arguments{
\item{url}{MPMP url.}
}
\description{
Hidden function to fetch the expandable tables from MPMP database
}
\keyword{internal}
