% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenfunctions.R
\name{.fast_ifelse2}
\alias{.fast_ifelse2}
\title{Fastest IfElse function: https://github.com/ICJIA/r-user-group/issues/11}
\usage{
.fast_ifelse2(test, yes, no)
}
\description{
Fastest IfElse function: https://github.com/ICJIA/r-user-group/issues/11
}
\keyword{internal}
