% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenfunctions.R
\name{.nonclickable}
\alias{.nonclickable}
\title{plasmoRUtils}
\usage{
.nonclickable(url)
}
\arguments{
\item{url}{MPMP url.}
}
\description{
Hidden function to download static tables from MPMP database
}
\keyword{internal}
