% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenfunctions.R
\name{.read_authenticated_tsv}
\alias{.read_authenticated_tsv}
\title{plasmoRUtils}
\usage{
.read_authenticated_tsv(url, api_key)
}
\description{
Hidden function to get tables from VEuPathDB databases with authentication
}
\keyword{internal}
