% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenfunctions.R
\name{.rs_api_search_page}
\alias{.rs_api_search_page}
\title{plasmoRUtils}
\usage{
.rs_api_search_page(
  query,
  offset = 0,
  article_type = "Research Article",
  status = "all"
)
}
\description{
Hidden function for searchRS
}
\keyword{internal}
