% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyGOPlot.R
\name{easyGOPlot}
\alias{easyGOPlot}
\title{Quick ORA result plotting}
\usage{
easyGOPlot(
  res,
  title = "GO Biological processes",
  limit = 20,
  desc = "Term",
  genecounts = "Annotated",
  sortby = "ks"
)
}
\arguments{
\item{res}{Output of \code{\link[=easytopGO]{easytopGO()}}.}

\item{title}{Title of the plot.}

\item{limit}{No of terms to plot.}

\item{desc}{Column name containing the GO description.If using VEuPathDB compnent database, this would be "Name" column.}

\item{genecounts}{Column name containing number of query genes associated with GO description. Used for bubble size. For VEupathDB this would be "Result count" column.}

\item{sortby}{Name of the statistics column to sort the terms by and use for plotting.}
}
\value{
A publication ready ggplot2 object.
}
\description{
A convenience function to quickly plot the results of \code{\link[=easytopGO]{easytopGO()}} or ORA results obtained from VEupathDB component databases.
}
\examples{
\dontrun{
## get enrichment results from easytopGO
gores <- easytopGO(geneID = geneList, bkggset = background.gset, stats = "ks")
baseurl <- "https://plasmodb.org/common/downloads/Current_Release/"
url <- paste0(baseurl, "Pfalciparum3D7/gaf/PlasmoDB-68_Pfalciparum3D7_GO.gaf.gz")
gores <- easytopGO(
  geneID = geneList, useGAF = TRUE, useBiomart = FALSE, gaf = url,
  bkggset = background.gset, category = "BP", stats = "ks"
)

## Then feed them to easyGOPlot
plot <- easyGOPlot(gores, title = "GO Enrichment Biological Processes")
}

}
