% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyMCA.R
\name{easyMCA}
\alias{easyMCA}
\title{Fetch data from Malaria Cell Atlas}
\usage{
easyMCA(url, type = "data")
}
\arguments{
\item{url}{A url of dataset from \code{\link[=listMCA]{listMCA()}} function.}

\item{type}{Type of data to be fetched. Use "exp" to fetch normalized and scaled values, use "raw" to get raw counts and use "data" to get the metadata of the dataset.}
}
\value{
df A dataframe of requested data type from MCA.
}
\description{
This function fetches and loads the metadata and expression matrices from the desired data sets available on Malaria Cell Atlas (MCA) Database.
}
\examples{
\dontrun{
  url <- "https://www.malariacellatlas.org/downloads/pf-ch10x-set4-biorxiv.zip"
  # Use the function to read metadata, expression, or raw data
  metadata <- easyMCA(url, type = "data")
  expression <- easyMCA(url, type = "exp")
  raw_counts <- easyMCA(url, type = "raw")

  ## make Seurat Object easily now

  testmca <- Seurat::CreateSeuratObject(counts = raw_counts, meta.data = metadata, project = "MCA")
}

}
