% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyPhplplottbl.R
\name{easyPhplplottbl}
\alias{easyPhplplottbl}
\title{Quickly plot PhenoPlasm summary tables}
\usage{
easyPhplplottbl(file, skip = 2)
}
\arguments{
\item{file}{Path of Phenotype.txt file obtained from PhenoPlasm database or a data frame.}

\item{skip}{Number of lines to skip in the file. Default 2.}
}
\value{
A gt table plot.
}
\description{
This function generates Disruptability and Mutant Phenotype tables in R, mirroring the style of Phenoplasm visualizations.
}
\examples{
\dontrun{

## Read the table generated from Phenoplasm advance search and pass the resulting data frame
df <- read.csv("phenotype.txt", skip = 2, sep = "\t") \%>\%
dplyr::select(-3, -4) \%>\%
dplyr::rename_with(~ gsub("Sprozoite", "Sporozoite", .x))
easyPhplplottbl(df)

## Pass the file path directly
easyPhplplottbl("phenotype.txt")

## Load example data frame
data(pf3d7PhplTable)

easyPhplplottbl(pf3d7PhplTable)
}

}
