% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyPie.R
\name{easyPie}
\alias{easyPie}
\title{Pie chart to summarize searchMT results}
\usage{
easyPie(df, col = "Tissue Specificity")
}
\arguments{
\item{df}{A dataframe obtained from plotTissueSpecific(returnData=TRUE).}

\item{col}{Column to plot as donut chart. Default: "Tissue Specificity"}
}
\value{
A plot (or data thereof) of domains present in the list of gene IDs.
}
\description{
This function make a donut chart to represent distribution of input gene IDs across different stages of Malaria given a result object from \code{plotTissueSpecific} function.
}
\examples{
\dontrun{
  geneID <- c("PBANKA_0100600", "PBANKA_0102900", "PF3D7_0102900")
  ## To get Plot similar to malaria.tools
  res <- searchMT(geneID = geneID)
  res \%>\% easyPie()
}

}
