% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyTPM.R
\name{easyTPM}
\alias{easyTPM}
\title{Calculate TPM values from count data}
\usage{
easyTPM(counts, featureLength)
}
\arguments{
\item{counts}{Count matrix containing raw counts.}

\item{featureLength}{Effective length of genes generated from \code{\link[=getEffLen]{getEffLen()}}.}
}
\value{
df numeric. This function returns a dataframe of TPM normalized counts and final column containing feature length.
}
\description{
This function provides ability to compute quick TPM values.
}
\examples{
\dontrun{
## Effective length of the gene
gene_info <- data.frame(GeneID = c("gene1", "gene2", "gene3"), Length = c(1000, 1500, 2000))
count_matrix <- matrix(c(10, 20, 30, 40, 50, 60),
  nrow = 3, ncol = 2,
  dimnames = list(c("gene1", "gene2", "gene3"), c("sample1", "sample2"))
)
test <- easyTPM(count_matrix, gene_info)
}

}
