% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEffLen.R
\name{getEffLen}
\alias{getEffLen}
\title{Get effective gene lengths form TPM calculation}
\usage{
getEffLen(gtf = NULL, format = "gff3")
}
\arguments{
\item{gtf}{Provide path or URL to GTF file.}

\item{format}{Format of the feature file i.e. "gtf" or "gff3". Default "gff3".}
}
\value{
df numeric. This function returns a data frame with 2 columns: "GeneID", "Length".
}
\description{
This function provides effective length (sum of lengths of exons) of the genes for calculating TPM values.
}
\examples{
\dontrun{

getEffLen("/data/PlasmoDB-67_Pfalciparum3D7.gtf")

}
}
