% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMpmp.R
\name{getMpmp}
\alias{getMpmp}
\title{Quickly fetch data from MPMP database}
\usage{
getMpmp(url)
}
\arguments{
\item{url}{URL of the pathway of interest.}
}
\value{
df This function returns a dataframe of containing the gene ID and Annotations fetched from MPMP database.
}
\description{
This function provides ability to fetch pathways data from MPMP. This data can then be modified and used for MPMP pathway enrichment analysis.
}
\examples{
\dontrun{
df <- getMpmp("http://mpmp.huji.ac.il/maps/HNE_prot.html")
df <- getMpmp("http://mpmp.huji.ac.il/maps/14-3-3prot.html")
}

}
