% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPreconfiguredTableOrthomcl.R
\name{getPreconfiguredTableOrthomcl}
\alias{getPreconfiguredTableOrthomcl}
\title{Get pre-configured tables from OrthoMCL-DB}
\usage{
getPreconfiguredTableOrthomcl(
  og_ids = NULL,
  customField = "previousGroups",
  timeout = 60,
  max_tries = 3
)
}
\arguments{
\item{og_ids}{Orthogroup IDs.}

\item{customField}{Preconfigured tables to be downloaded from OrthoMCL.}

\item{timeout}{Timeout time in seconds}

\item{max_tries}{Maximum retries allowed.

\itemize{
\item \emph{"TaxonCounts"}: Phlyetic Distribution of Proteins
\item \emph{"PhyleticDistributionCounts"}: Phyletic Distribution clade abundance
\item \emph{"previousGroups"}: All previous groups."Default"
\item \emph{"GroupStat"}: Group Statistics
\item \emph{"SimilarGroup"}: Similar Groups
\item \emph{"EcNumber"}: Summary of EC Numbers
\item \emph{"PFams"}: Summary of Pfam domains
\item \emph{"ProteinPFams"}: PFam Architecture of Each Protein
\item \emph{"KeywordFrequency"}: Keyword Frequency
}}
}
\value{
A data frame.
}
\description{
A convenience function to quickly fetch preconfigured table of from OrthoMCL-DB. This function can be used to map old OG IDs to new and vice-versa given a vector of OG IDs. If no IDs are provided, entire data tables fetched from the database will be returned, which can be manipulated or processed further as per user need.
}
\examples{
\dontrun{
df <- getPreconfiguredTableOrthomcl("OG3_10277")
df <- getPreconfiguredTableOrthomcl(customField="SimilarGroup")
}

}
