% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTable}
\alias{getTable}
\title{Get tables with custom fields}
\usage{
getTable(
  org,
  db = "toxodb",
  customFields = NULL,
  api_key = Sys.getenv("VEUPATHDB_API_KEY")
)
}
\arguments{
\item{org}{Full name of organism of interest as specified in VEuPathDB. To find the exact name of the organism, use \code{listVeupathdb} function.}

\item{db}{Character Name of the database in which the organism is present. These can be one of the following: "toxodb","plasmodb","hostdb","amoebadb","cryptodb","fungidb","giardiadb","microsporidiadb","piroplasmadb","trichdb","tritrypdb".}

\item{customFields}{A vector of custom fields desired to be fetched. "primary_key" is mandatory field. Use \code{\link[=getTableAttributes]{getTableAttributes()}} to browse the attributes available for the
selected database and record type. If \code{NULL}, a default set of
attributes is returned.}

\item{api_key}{VEuPathDB API key.}
}
\value{
A data frame, containing "Gene ID", "Product Description", "Gene Strand", "Gene Name or Symbol", "Previous ID(s)", "Entrez Gene ID", "UniProt ID(s)", "Protein Length", "TM Domains" and "SignalP Peptide" for all the genes present in the organism of interest.
}
\description{
A convenience function to quickly fetch table of Gene IDs, Protein IDs, Gene Symbols, Annotations and many more columns from database of your choice such as PlasmoDB, ToxoDB, PiroplasmaDB among other VEuPathDB pathogen databases.
}
\details{
With release 71 of VEuPathDB, only users with subscription and API key can access the database programmatically. Use \code{usethis::edit_r_environ(scope = "user")} to add VEuPathDB API using \code{VEUPATHDB_API_KEY} variable.
}
\examples{
\dontrun{

df <- getTable(org="Plasmodium falciparum 3D7", 
db="plasmodb",api_key = Sys.getenv("VEUPATHDB_API_KEY"))

df <- getTable(org="Plasmodium falciparum 3D7", 
db="plasmodb", api_key = Sys.getenv("VEUPATHDB_API_KEY"), 
customFields = c("primary_key","protein_sequence"))
}

}
\seealso{
\code{\link[=getTableAttributes]{getTableAttributes()}} for browsing the attributes available
for a given record type.
}
