% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableAttributes.R
\name{getTableAttributes}
\alias{getTableAttributes}
\title{Browse available VEuPathDB databases attributes}
\usage{
getTableAttributes(
  db = "plasmodb",
  record_type = "transcript",
  api_key = Sys.getenv("VEUPATHDB_API_KEY")
)
}
\arguments{
\item{db}{Character. Name of the VEuPathDB database, such as
\code{"plasmodb"}, \code{"toxodb"}, \code{"fungidb"} etc.}

\item{record_type}{Character. VEuPathDB record type for which attributes
should be retrieved. Defaults to \code{"transcript"}.}

\item{api_key}{Character. VEuPathDB API key used for authentication.
By default, the value is obtained from the \code{VEUPATHDB_API_KEY}
environment variable.}
}
\value{
A data frame containing metadata for the available attributes,
including the internal attribute name, display name, data type,
reporting availability, and help text where available.
}
\description{
Retrieves metadata describing the attributes available for a specified
VEuPathDB record type. The returned attribute names can be supplied to
the \code{customFields} argument of \code{\link[=getTable]{getTable()}}.
}
\details{
With release 71 of VEuPathDB, only users with subscription and API key can access the database programmatically. Use \code{usethis::edit_r_environ(scope = "user")} to add VEuPathDB API using \code{VEUPATHDB_API_KEY} variable.

VEuPathDB record types can contain many reportable attributes.
\code{getAttributes()} provides a convenient way to browse these attributes
without manually consulting the VEuPathDB web service.

The \code{name} column contains the attribute identifiers expected by the
\code{customFields} argument of \code{\link[=getTable]{getTable()}}, while \code{displayName} provides
human-readable labels.
}
\examples{
\dontrun{
# Browse transcript attributes available in PlasmoDB
attrs <- getTableAttributes(
  db = "plasmodb",
  record_type = "transcript", api_key=Sys.getenv("VEUPATHDB_API_KEY")
)

head(attrs)

# Search for protein-related attributes
subset(
  attrs,
  grepl("protein", displayName, ignore.case = TRUE)
)

# Use selected attribute names with getTable()
genes <- getTableAttributes(
  org = "Plasmodium falciparum 3D7",
  db = "plasmodb",
  customFields = c(
    "primary_key",
    "gene_product",
    "protein_length"
  )
)
}

}
\seealso{
\code{\link[=getTable]{getTable()}} for retrieving data using selected attributes.
}
\concept{VEuPathDB data retrieval}
