% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dependencies.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Install package dependencies}
\usage{
install_dependencies()
}
\description{
Installs missing dependencies from both CRAN and Bioconductor
}
