% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listipdb.R
\name{listipdb}
\alias{listipdb}
\title{List species metadata present in InparanoiDB 9}
\usage{
listipdb()
}
\value{
A data frame containing information about species present in InParanoiDB9 and their taxon ID.
}
\description{
A convenience function to quickly fetch table of all InParanoiDB 9 species and their taxonomy id.
}
\examples{
\dontrun{
listipdb()
}

}
