% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOrthomcl.R
\name{listOrthomcl}
\alias{listOrthomcl}
\title{List species metadata present in OrthoMCL}
\usage{
listOrthomcl(api_key = Sys.getenv("VEUPATHDB_API_KEY"))
}
\arguments{
\item{api_key}{Character. VEuPathDB API key used for authentication.
By default, the value is obtained from the \code{VEUPATHDB_API_KEY}
environment variable.}
}
\value{
A data frame containing information about species present in OrthoMCL and their database associated ID.
}
\description{
A convenience function to quickly fetch the species related vocabulary of ≥ 857 species used by \href{https://orthomcl.org/orthomcl/app}{OrthoMCL} database. This function helps users choose numeric IDs of the organisms between which they wish to fetch the paired orthologs. See also: \code{\link[=getpairedOrthologs]{getpairedOrthologs()}}.
}
\details{
User should note that the numeric IDs used by OrthoMCL are not stable and changes upon every update. So an ID which was previously assigned to \verb{Plasmodium falciparum 3D7} might get assigned to a new organism in new release. Therefore users are encouraged to run this function prior to running of \code{\link[=getpairedOrthologs]{getpairedOrthologs()}}
}
\examples{
\dontrun{
listOrthomcl()
}

}
