% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listVeupathdb.R
\name{listVeupathdb}
\alias{listVeupathdb}
\title{List genomes and metadata in VEuPathDB}
\usage{
listVeupathdb(
  customFields = NULL,
  api_key = Sys.getenv("VEUPATHDB_API_KEY"),
  timeout = 60,
  max_tries = 3
)
}
\arguments{
\item{customFields}{A vector of custom fields desired to be fetched. "primary_key" is mandatory field. Other fields can be supplied and can be chosen from (but are not limited to): \code{"annotation_source"}, \code{"annotation_version"}, \code{"arraygenecount"}, \code{"chipchipgenecount"}, \code{"chromosomeCount"}, \code{"codinggenecount"}, \code{"communitycount"}, \code{"contigCount"}, \code{"ecnumbercount"}, \code{"estcount"}, \code{"genecount"}, \code{"genecount_number"}, \code{"genome_source"}, \code{"genome_version"}, \code{"gocount"}, \code{"is_in_apollo"}, \code{"is_reference_strain"}, \code{"megabps"}, \code{"ncbi_tax_id"}, \code{"ncbi_taxon_url"}, \code{"organism"}, \code{"organism_full"}, \code{"orthologcount"}, \code{"othergenecount"}, \code{"popsetcount"}, \code{"project_id"}, \code{"proteomicscount"}, \code{"pseudogenecount"}, \code{"rnaseqcount"}, \code{"rtpcrcount"}, \code{"snpcount"}, \code{"species"}, \code{"species_ncbi_tax_id"}, \code{"species_ncbi_taxon_url"}, \code{"supercontigCount"}, \code{"tfbscount"}, \code{"URLcdsFasta"}, \code{"URLGenomeFasta"}, \code{"URLgff"}, \code{"URLproteinFasta"}, \code{"URLtranscriptFasta"}. For more fields, refer to the VEuPathDB Documentation}

\item{api_key}{Character. VEuPathDB API key used for authentication.
By default, the value is obtained from the \code{VEUPATHDB_API_KEY}
environment variable.}

\item{timeout}{Timeout time in seconds. Defaults to 60.}

\item{max_tries}{Maximum number of request attempts. Defaults to 3.}
}
\value{
A data frame containing information about genomes present in VEuPathDB and their attributes.
}
\description{
A convenience function to quickly fetch table of genomes and their associated metadata from VEupathDB.
}
\examples{
\dontrun{
df <- listVeupathdb()
df <- listVeupathdb(customFields=c("species", "project_id"))
}

}
