% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb2uniprot.R
\name{pdb2uniprot}
\alias{pdb2uniprot}
\title{Quick mapping of PDB IDs to Uiprot IDs}
\usage{
pdb2uniprot(pdbid, timeout = 60, max_tries = 3)
}
\arguments{
\item{pdbid}{A single PDB Id.}

\item{timeout}{Timeout time in seconds}

\item{max_tries}{Maximum retries allowed.}
}
\value{
A data frame, Uniprot IDs, Chain IDs and start and end coordinates of the chains.
}
\description{
A convenience function to quickly convert the PDB IDs to Uniprot IDs. If a protein is multimeric, corresponding Uniprot IDs are returned for them. This function uses PDBe API.
}
\examples{
\dontrun{
df <- pdb2uniprot("9FIA")
}

}
