% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchApicoTFdb.R
\name{searchApicoTFdb}
\alias{searchApicoTFdb}
\title{Fetch data from ApicoTFdb}
\usage{
searchApicoTFdb(org = "pf")
}
\arguments{
\item{org}{Abbreviation of organism of interest.
Plasmodium Species:
pb: Plasmodium berghii
pv: Plasmodium vivax
pf: Plasmodium falciparum
pk: Plasmodium knowlesi
py: Plasmodium yoelii
pc: Plasmodium chabaudi

By Other Apicomplexan Species
tg49: Toxoplasma Gondii ME49
tg89: Toxoplasma Gondii P89
cp: Cryptosporidium parvum
em: Eimeria maxima
bb: Babesia bovis
et: Eimeria tenella
nu: Neurospora caninum
cy: Cyclospora cayetanensis}
}
\value{
df This function returns a dataframe of transcription regulators for organism of interest from ApicoTFDb.
}
\description{
This function provides ability to query your gene IDs to ApicoTFDb.
}
\examples{
\dontrun{
test <- searchApicoTFdb(org="pf")
}

}
