% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchGSC.R
\name{searchGSC}
\alias{searchGSC}
\title{Fetch articles for given gene IDs from Google Scholar}
\usage{
searchGSC(geneID)
}
\arguments{
\item{geneID}{Character vector of Gene IDs. If you want to use gene symbols, use organism name alongside to avoid articles that might have similar abbreviated word. eg. use "AP2-P AND Plasmodium".}
}
\value{
A data frame, containing 5 columns: GeneID, Title of the article, Year of Publication, Url and Authors.
}
\description{
This function searches the Google Scholar corpus for the articles that contains your Gene ID of interest.
}
\details{
\strong{Warning}: Scraping Google Scholar is against their \href{https://scholar.google.com/intl/en/scholar/about.html}{Terms of Service}. We advise users to use this function for querying few IDs (not more than 20) per day.
Proceeding with this function may result in your IP being blocked temporarily.
}
\examples{
\dontrun{
test <- searchGSC(c("PF3D7_0420300", "PF3D7_0621000"))
test <- searchGSC(geneID = c("AP2-P AND Plasmodium","AP2-I"))
}

}
