% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchKipho.R
\name{searchKipho}
\alias{searchKipho}
\title{Fetch Kinases and Phosphatases from Kipho Database}
\usage{
searchKipho(org = "pf", type = "kinase")
}
\arguments{
\item{org}{Abbreviation of organism of interest.
Plasmodium Species:
pb: Plasmodium berghii
pv: Plasmodium vivax
pf: Plasmodium falciparum
pc: Plasmodium chabaudi}

\item{type}{Type of protein class i.e. "kinase" or "phosphatase". Default: "kinase"}
}
\value{
df This function returns a dataframe of kinases/phosphatases in Plasmodium species.
}
\description{
This function provides ability to query your gene IDs to KiPho Database.
}
\examples{
\dontrun{
test <- searchKipho(org="pf")
}

}
