% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMidb.R
\name{searchMidb}
\alias{searchMidb}
\title{Fetch Minor intron tables from MiDB}
\usage{
searchMidb(org, type = "intron")
}
\arguments{
\item{org}{Name of Organism. Can be obtained by loading midbSpecies data from the package}

\item{type}{Type of data to be fetched. Default: "intron"}
}
\value{
df This function returns a dataframe of intron classification in Plasmodium species.
}
\description{
This function provides ability to fetch the intron class data for 265 species in MiDB database. For more information refer to the \href{https://midb.pnb.uconn.edu/aboutus.php}{MiDB database}
}
\examples{
\dontrun{
load("data/midbSpecies.rda")
## Fetching intron data from MiDB for P. falciparum
df <- searchMidb(midbSpecies$`Available Species`[196])
}

}
