% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMT.R
\name{searchMT}
\alias{searchMT}
\title{Fetch data tables from malaria.tools database}
\usage{
searchMT(geneID)
}
\arguments{
\item{geneID}{A character vector of Gene IDs of Plasmodium falciparum or Plasmodium berghi.}
}
\value{
A plot (or data thereof) of TPM values across multiple stages of parasite.
}
\description{
This function retrieves data from malaria.tools and generates a dataframe containing the Stage of Parasite in which the gene is highly expressed.
}
\examples{
\dontrun{
  geneID <- c("PBANKA_0100600", "PBANKA_0102900", "PF3D7_0102900")
  ## To get condition specificity and tissue specificity data
  res <- searchMT(geneID = geneID)
}

}
