% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchPhPl.R
\name{searchPhPl}
\alias{searchPhPl}
\title{Fetch data tables from PhenoPlasm database}
\usage{
searchPhPl(geneID = "", org = "pf", fetch = 1)
}
\arguments{
\item{geneID}{Character vector of Gene IDs.}

\item{org}{Abbreviation of the organism. Default "pf"}

\item{fetch}{Numeric. Use 1 to fetch the "Disruptability" table and 2 to fetch "Mutant phenotypes" table.
\itemize{
\emph{Plasmodium Species:}
\item pb: \emph{Plasmodium berghii}
\item pk: \emph{Plasmodium knowlesi}
\item pf: \emph{Plasmodium falciparum}
\item pc: \emph{Plasmodium chabaudi}
\item py: \emph{Plasmodium yoelii}
}}
}
\value{
A data frame.
}
\description{
This function searches the Phenotypes of the gene IDs in Phenoplasm database and enables users to fetch sub-tables such as Disruptability and Mutant phenotypes.
}
\examples{
\dontrun{
## get phenotype for few genes in plasmodium falciparum
df <- searchPhPl(geneID = c("PF3D7_0420300","PF3D7_0621000","PF3D7_0523800"), org="pf")
df <- searchPhPl(geneID = c("PF3D7_0420300","PF3D7_0621000","PF3D7_0523800"), org="pf", fetch=2)

}

}
