% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTedConsensus.R
\name{searchTedConsensus}
\alias{searchTedConsensus}
\title{Fetch protein domains from TED database}
\usage{
searchTedConsensus(
  uniprotid = "",
  returnCATHdesc = TRUE,
  timeout = 60,
  max_tries = 3
)
}
\arguments{
\item{uniprotid}{A character vector of uniprot IDs.}

\item{returnCATHdesc}{Logical. Set this on to get the description of the CATH ID from CATH database.}

\item{timeout}{Timeout time in seconds}

\item{max_tries}{Maximum retries allowed.}
}
\value{
A data frame, domain boundaries and other information provided by TED. For details visit the TED database.
}
\description{
A convenience function to quickly access The Encyclopedia of Domains (TED) database and fetch domain boundary information for given Uniprot IDs. For information about the column names the users are requested to refer to the TED database at https://ted.cathdb.info/.
}
\examples{
\dontrun{
df <- searchTedConsensus(
c("Q7K6A1","Q8IAP8","C0H4D0","C6KT90","Q8IBJ7"),
returnCATHdesc=FALSE)
}

}
