% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toGeneid.R
\name{toGeneid}
\alias{toGeneid}
\title{Convert Other IDs to Ensembl gene IDs}
\usage{
toGeneid(
  inputid,
  from = "",
  to = "",
  org = "Plasmodium falciparum 3D7",
  db = "plasmodb",
  ...
)
}
\arguments{
\item{inputid}{A character vector of IDs. Can be Ensembl, Uniprot, Entrez or or old Pf ids.}

\item{from}{To describe the type of Input ID. Possible values: "old", "uniprot". "entrez", "ensembl"}

\item{to}{To describle the type of output ID desired. Possible values: "emsembl".}

\item{org}{Organism to which the IDs belongs to.Possible values: toxodb, plasmodb, hostdb, amoebadb, cryptodb, fungidb, giardiadb, microsporidiadb, piroplasmadb, trichdb, tritrypdb.}

\item{db}{Database in which organism is present.}

\item{...}{Additional arguments that can be passed to the \code{\link[=getTable]{getTable()}} function.}
}
\value{
A data frame, containing Gene IDs, gene description and gene Symbols and more.
}
\description{
A convenience function to quickly convert the Uniprot or Entrez Ids to Ensembl gene IDs,  using VEuPathDB specialized databases.It also provides description and gene symbol for input Ids.
}
\details{
With release 71 of VEuPathDB, only users with subscription and API key can access the database programmatically. Use \code{usethis::edit_r_environ(scope = "user")} to add VEuPathDB API using \code{VEUPATHDB_API_KEY} variable.
}
\examples{
\dontrun{
df <- toGeneid(
c("PF3D7_0420300", "PF3D7_0621000"),
      from="ensembl",api_key=Sys.getenv("VEUPATHDB_API_KEY"))
}

}
