% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenfunctions.R
\name{.usetopGO}
\alias{.usetopGO}
\title{plasmoRUtils}
\usage{
.usetopGO(
  stats = stats,
  category = category,
  geneID = geneID,
  gene_2_GO = gene_2_GO,
  algo = algo,
  fdr = fdr,
  correction = correction
)
}
\description{
Hidden function called by \code{easytopGO} to run topGO.
}
\keyword{internal}
