% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPreconfiguredTableOrthomcl.R
\name{getPreconfiguredTableOrthomcl}
\alias{getPreconfiguredTableOrthomcl}
\title{Get pre-configured tables from OrthoMCL-DB}
\usage{
getPreconfiguredTableOrthomcl(
  og_ids = NULL,
  customField = "previousGroups",
  api_key = Sys.getenv("VEUPATHDB_API_KEY"),
  timeout = 60,
  max_tries = 3,
  refresh = FALSE
)
}
\arguments{
\item{og_ids}{Orthogroup IDs.}

\item{customField}{Preconfigured table to be downloaded from OrthoMCL.}

\item{api_key}{Character. VEuPathDB API key used for authentication.
By default, the value is obtained from the \code{VEUPATHDB_API_KEY}
environment variable.}

\item{timeout}{Timeout time in seconds.}

\item{max_tries}{Maximum retries allowed.}

\item{refresh}{Logical. If \code{TRUE}, ignores any locally cached copy and
downloads the requested table again from OrthoMCL. Defaults to \code{FALSE}.

\itemize{
\item \emph{"TaxonCounts"}: Phyletic Distribution of Proteins
\item \emph{"PhyleticDistributionCounts"}: Phyletic Distribution clade abundance
\item \emph{"previousGroups"}: All previous groups. "Default"
\item \emph{"GroupStat"}: Group Statistics
\item \emph{"SimilarGroup"}: Similar Groups
\item \emph{"EcNumber"}: Summary of EC Numbers
\item \emph{"PFams"}: Summary of Pfam domains
\item \emph{"ProteinPFams"}: PFam Architecture of Each Protein
\item \emph{"KeywordFrequency"}: Keyword Frequency
}}
}
\value{
A data frame.
}
\description{
A convenience function to quickly fetch preconfigured tables from
OrthoMCL-DB. This function can be used to map old OG IDs to new and
vice-versa given a vector of OG IDs. If no IDs are provided, the entire
data table fetched from the database is returned and a copy of it is cached to
reduce the next query time.
}
\examples{
\dontrun{
df <- getPreconfiguredTableOrthomcl("OG3_10277")
df <- getPreconfiguredTableOrthomcl(customField = "SimilarGroup")
}

}
