% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMCA.R
\name{listMCA}
\alias{listMCA}
\title{List data sets in MCA}
\usage{
listMCA()
}
\value{
df A dataframe listing all the datasets available on MCA and URls that can be used with \code{easyMCA} function.
}
\description{
This function lists all the datasets available on Malaria Cell Atlas Database.
}
\examples{
\dontrun{
  df <- listMCA()
}

}
