% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOrthomcl.R
\name{listOrthomcl}
\alias{listOrthomcl}
\title{List species metadata present in OrthoMCL}
\usage{
listOrthomcl()
}
\value{
A data frame containing information about species present in InParanoiDB9 and their taxon ID.
}
\description{
A convenience function to quickly fetch the species related vocabulary of 713 species used by \href{https://orthomcl.org/orthomcl/app}{OrthoMCL} database. This function helps users choose IDs of the organisms between which they wish to fetch the paired orthologs. See also: \code{getpairedOrthologs()}.
}
\examples{
\dontrun{
listOrthomcl()
}

}
