% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchIpDb.R
\name{searchIpDb}
\alias{searchIpDb}
\title{Fetch orthologs from InParanoiDb}
\usage{
searchIpDb(geneID, ...)
}
\arguments{
\item{geneID}{Gene ID of Plasmodium falciparum.}

\item{...}{Additional arguments that can be passed to \code{toGeneid} function.}
}
\description{
This function retrieves the Plasmodium gene orthologs from InParanoiDB 9 database across all life forms
}
\examples{
\dontrun{
ids <- c("PF3D7_0807800", "PF3D7_1023900")
df <- searchIpDb(ids)
}

}
