% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchIpDb.R
\name{searchIpDb}
\alias{searchIpDb}
\title{Fetch orthologs from InParanoiDb}
\usage{
searchIpDb(geneID, ..., idtype = "ensembl")
}
\arguments{
\item{geneID}{Gene ID of \emph{Plasmodium falciparum}.}

\item{...}{Additional arguments that can be passed to \code{toGeneid()} function. This comes in very handy if you are working with parasite gene IDs other than \emph{Plasmodium}.

@return A data frame, containing 10 columns.}

\item{idtype}{Set this to "uniprot" if using uniprot IDs directly and GenID to uniprot ID conversion is not required.}
}
\description{
This function retrieves the orthologs from \href{https://inparanoidb.sbc.su.se/}{InParanoiDB 9} database for 640 species given set of Ensembl Gene IDs or uniprot IDs usinf database API.
}
\details{
To view list of species covered by InParanoiDB 9, use \code{listipdb()} function.
}
\examples{
\dontrun{
df <-  searchIpDb( c("PF3D7_0807800", "PF3D7_1023900"))
df <- searchIpDb( c("C5LD32", "A5KAC7"),idtype = "uniprot")
}

}
\seealso{
\code{\link[plasmoRUtils]{listipdb}}
}
