% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMiip.R
\name{searchMiip}
\alias{searchMiip}
\title{Fetch Protein-Protein interactions from MIIP database}
\usage{
searchMiip(geneID)
}
\arguments{
\item{geneID}{A character vector of Gene IDs of Plasmodium falciparum or Plasmodium berghi.Remove version from the gene ids.}
}
\value{
A dataframe of Protein protein interaction provided by MIIP database.
}
\description{
This function retrieves Protein-protein interaction data from \href{http://www.hpppi.iicb.res.in/pfnet/}{MIIP database}.
}
\examples{
\dontrun{
 df <- searchMiip(c("PF3D7_0807800","PF3D7_1023900"))
}

}
