% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchPM.R
\name{searchPM}
\alias{searchPM}
\title{Fetch articles from PubMed database}
\usage{
searchPM(
  geneID,
  org = "Plasmodium falciparum",
  query = NULL,
  from = 2010,
  to = 2025,
  verbose = TRUE
)
}
\arguments{
\item{geneID}{Character vector of Gene IDs.}

\item{org}{Scientific name of the organism. Default \emph{Plasmodium falciparum}}

\item{query}{String of user defined custom queries. If you wish to pass your own combination od terms use this argument.}

\item{from}{To define the start year for querying the articles.}

\item{to}{To define the end year for querying the articles.}

\item{verbose}{Disable to turn off the messages printed by the function.}
}
\value{
A data frame, containing 9 columns: "pmid"    "doi"     "title"   "year"    "month"   "day"     "jabbrv"  "journal" "GeneID" .
}
\description{
This function searches the PubMed corpus for the articles that contains your Gene ID of interest.
}
\examples{
\dontrun{
test <- searchPM(geneID = c("PF3D7_0420300","PF3D7_0621000"))
}

}
